Refactor SubstrateIdentity out of User widget.
