+++ client/scripts/views/components/header/login_selector.ts
@@ -16,10 +16,7 @@ import { notifySuccess } from 'controllers/app/notifications';
 import { ChainIcon, CommunityIcon } from 'views/components/chain_icon';
 import ChainStatusIndicator from 'views/components/chain_status_indicator';
 import User, { UserBlock } from 'views/components/widgets/user';
-import LinkNewAddressModal from 'views/modals/link_new_address_modal';
 import LoginModal from 'views/modals/login_modal';
-import EditIdentityModal from 'views/modals/edit_identity_modal';
-import EditProfileModal from 'views/modals/edit_profile_modal';
 import FeedbackModal from 'views/modals/feedback_modal';
 import SelectAddressModal from 'views/modals/select_address_modal';
 import { setActiveAccount } from 'controllers/app/login';
